data_collector/requirements.txt
pytrends
pandas
psycopg2-binary
python-dotenv
bs4